+++ src/Model/ProductInterface.php
@@ -36,7 +36,7 @@ interface ProductInterface
 
     public function getName(): string;
 
-    public function getPrice(): float;
+    public function getPrice(): int;
 
     public function getPoints(): int;
